+++ src/app/app.module.ts
@@ -9,7 +9,7 @@ import { AwsModule } from '../domain/aws/aws.module';
 import { GameModule } from '../domain/game/game.module';
 import { BullModule } from '@nestjs/bull';
 import { RedisModule } from '@nestjs-modules/ioredis';
-import { bullqConfig, redisConfig, typeormConfig } from '~/src/app/config';
+import { bullqConfig, redisConfig, typeormConfig } from '~/src/app/configs/redis.config';
 
 @Module({
   imports: [
