+++ src/app/configs/bullq.config.ts
@@ -1,0 +1,8 @@
+import { ConfigService } from '@nestjs/config';
+export const bullqConfig = (configService: ConfigService) => ({
+  redis: {
+    host: configService.get<string>('REDIS_HOST', 'localhost'),
+    port: configService.get<number>('REDIS_PORT', 6379),
+    password: configService.get<string>('REDIS_PASSWORD'),
+  },
+});
